Allow fine grained event control
This change enables the ability to remove/add events that should be
handled by the eventHandler function. The purpose for this change is to
fix the cursor issue that occurs with Android browsers.

If you remove the 'input' event from the eventsToHandle then the ui-mask
directive works properly for Android browsers. However, you can't just
remove it permanently as that causes other UX issues on non-Android
browsers.