Fix nested in ng-if issue
Added a check to the initPlaceholder function to ensure it only updates
the element value if there is a value on the element or if there is not
a placeholder attribute on the element.

If the element value is blank and the element has a placeholder
attribute, then the element's value does not get updated.