Fix backspace issue
Removed the confusing RegEx replace operation in getMaskComponents that
didn't perform as intended and caused edge case issues in others

Added a keydown handler to detect if a backspace operation should be
allowed to happen. Backspace should occur properly when there is a valid
maskable character position to remove but should not happen if there are
only static mask characters.

Examples:

Mask = QT****
Value = QT____

A backspace is not valid in this case and would previously cause the Q
to leak into the value

Mask = QT****
Value = QT1___

A backspace is valid in this cause and will remove the 1.